osbuild: split `pipeline`
This splits out the long, monolithic `pipeline.py` into a `manifest`
subpackage. This allows multiple files which contain less code.
