+++ osbuild/formats/v1.py
@@ -11,7 +11,7 @@ from typing import Any, Dict
 
 from osbuild.meta import Index, ValidationResult
 
-from ..pipeline import BuildResult, Manifest, Pipeline, Runner
+from ..manifest import BuildResult, Manifest, Pipeline, Runner
 
 VERSION = "1"
